Improved build scripts and kalman_filter

+++ src/ground_truth_package.h
@@ -6,7 +6,7 @@
 class GroundTruthPackage
 {
   public:
-    long timestamp_;
+    long long timestamp_;
 
     enum SensorType
     {
